[modularize] Put files back because we have not decided the path of modules.
